// apps/admin/src/pages/Exhibition/Register/ExhibitionRegister.types.ts
export interface ArtistInfo {
  id: number;
  profileImage: string;
  name: string;
  major: string;
  email: string;
}

export interface ExhibitionDetailInfo {
  exhibitType: string;
  year: string;
  major: string;
  title: string;
  subTitle: string;
  description_en: string;
  description_ko: string;
  behanceUrl: string;
  linkedinUrl: string;
  youtubeUrl: string;
}